requirements.txt
urllib3<2
certifi==2024.2.2
polars==0.20.25
tqdm==4.66.4
spacy==3.7.4
pydantic==2.7.1
llama_cpp_python==0.2.75
scipy==1.13.1
matplotlib==3.9.0
instructor==1.2.6
nl-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/nl_core_news_sm-3.7.0/nl_core_news_sm-3.7.0-py3-none-any.whl#sha256=9255c3ce9e639f0a7aa8e898f4514e6ad803284858241385e045f99637753770
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889